requirements.txt
geopandas
fiona
shapely
pyproj
Office365-REST-Python-Client
requests
pendulum